deploy single page assistant app
