+++ genai/streamlit-assistant/requirements.txt
@@ -1,0 +1,3 @@
+streamlit 
+vertexai
+google-cloud-aiplatform
