--- genai/streamlit/requirements.txt
@@ -1,10 +1,0 @@
-streamlit 
-matplotlib
-networkx
-google-cloud-secret-manager
-numpy
-pandas
-duckdb
-vertexai
-google-cloud-storage
-google-cloud-aiplatform
